% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{cat0}
\alias{cat0}
\title{Generowanie ciągów tekstowych bez spacji}
\usage{
cat0(..., sep = "")
}
\arguments{
\item{...}{Obiekty, które mają zostać połączone.}

\item{sep}{Separator, który ma zostać użyty. Domyślnie jest to pusty ciąg znaków.}
}
\value{
zwraca ciąg tekstowy
}
\description{
Funkcja do łączenia tekstów bez spacji funkcja skopiowana z
https://rdrr.io/cran/rosetta/src/R/cat0.R
}
