% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych_D1.R
\name{dane_tab_D1_zaw}
\alias{dane_tab_D1_zaw}
\title{Przygotowanie danych do tabel}
\usage{
dane_tab_D1_zaw(
  pelna_finalna_ramka_wskaznikow,
  typ_szk,
  rok_absolwentow,
  tylko_tabele
)
}
\arguments{
\item{pelna_finalna_ramka_wskaznikow}{Ramka danych zawierająca pełne wyniki wskaźników.
Oczekuje, że kolumna 'wynik' zawiera zagnieżdżone ramki danych.}

\item{typ_szk}{Zmienna tekstowa opisująca typ szkoły}

\item{rok_absolwentow}{Liczba całkowita reprezentująca rok absolwentów do filtrowania.}

\item{tylko_tabele}{parametr TRUE/FALSE przekazywany z głównej funkcji
generującej raport - jeśli FALSE to przycina tabele z zawodami do 10
najliczniejszych zawodów, jeśli TRUE to raport generuje tabele zawodów bez
przycinania}
}
\value{
Ramka danych typu tibble w finalnym formacie do zasilania tabel.
}
\description{
Funkcja filtruje i przetwarza dane z pełnej ramki wskaźników,
aby przygotować je do generowania tabel.
}
