% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych_K1.R
\name{dane_tab_K1_zaw}
\alias{dane_tab_K1_zaw}
\title{Przygotowanie danych do wykresu statusów (płeć)}
\usage{
dane_tab_K1_zaw(
  pelna_finalna_ramka_wskaznikow,
  typ_szk,
  rok_absolwentow,
  rok,
  tylko_tabele
)
}
\arguments{
\item{pelna_finalna_ramka_wskaznikow}{Ramka danych zawierająca pełne wyniki wskaźników.
Oczekuje, że kolumna 'wynik' zawiera zagnieżdżone ramki danych.}

\item{typ_szk}{Zmienna tekstowa opisująca typ szkoły.}

\item{rok_absolwentow}{Liczba całkowita reprezentująca rok absolwentów do filtrowania.}

\item{rok}{Liczba całkowita reprezentująca rok do filtrowania.}

\item{tylko_tabele}{parametr TRUE/FALSE przekazywany z głównej funkcji
generującej raport - jeśli FALSE to przycina tabele z zawodami do 10
najliczniejszych zawodów, jeśli TRUE to raport generuje tabele zawodów bez
przycinania}
}
\value{
Ramka danych typu tibble w finalnym formacie do zasilania wykresów.
}
\description{
Funkcja wyciąga i przetwarza dane z pełnej ramki wskaźników,
aby przygotować je do generowania wykresu. Tworzy ramkę danych
z kolumnami \code{status}, \code{plec}, oraz \code{pct} (jako odsetek) w formacie
będącym wejściem do funkcji wykresStatusyPlec z pakietu LOSYkolory.
}
