% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych_meta.R
\name{dane_tab_meta_typsz_og}
\alias{dane_tab_meta_typsz_og}
\title{Przygotowanie danych do tabeli (typy szkół ogółem)}
\usage{
dane_tab_meta_typsz_og(pelna_finalna_ramka_wskaznikow, rok_absolwentow)
}
\arguments{
\item{pelna_finalna_ramka_wskaznikow}{Ramka danych zawierająca pełne wyniki wskaźników.
Oczekuje, że kolumna 'wynik' zawiera zagnieżdżone ramki danych.}

\item{rok_absolwentow}{Liczba całkowita reprezentująca rok absolwentów do filtrowania.}
}
\value{
Ramka danych typu tibble w finalnym formacie do zasilania wykresów i tabel.
}
\description{
Funkcja filtruje i przetwarza dane z pełnej ramki wskaźników,
aby przygotować je do generowania tabel. Filtruje dane dla całej Polski, dla wskaźnika
'typ_szk2'. Przekształca dane i oblicza procenty oraz liczebności
w formacie zgodnym z raportowaniem.
}
