% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych_S7.R
\name{dane_tab_S7_zaw}
\alias{dane_tab_S7_zaw}
\title{Przygotowanie danych do tabel}
\usage{
dane_tab_S7_zaw(pelna_finalna_ramka_wskaznikow, typ_szk, rok_absolwentow, rok)
}
\arguments{
\item{pelna_finalna_ramka_wskaznikow}{Ramka danych zawierająca pełne wyniki wskaźników.
Oczekuje, że kolumna 'wynik' zawiera zagnieżdżone ramki danych.}

\item{typ_szk}{Zmienna tekstowa opisująca typ szkoły}

\item{rok_absolwentow}{Liczba całkowita reprezentująca rok absolwentów do filtrowania.}

\item{rok}{Liczba całkowita reprezentująca rok do filtrowania.}
}
\value{
Ramka danych typu tibble w finalnym formacie do zasilania wykresów.
}
\description{
Funkcja filtruje i przetwarza dane z pełnej ramki wskaźników,
aby przygotować je do generowania wykresów. Tworzy ramkę danych
w formacie bedącym wejściem do funkcji gentab_tab_S7_zawod z tego pakietu.
}
