% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych_W1.R
\name{dane_tab_W1}
\alias{dane_tab_W1}
\title{Przygotowanie danych do tabeli wynagrodzeń (płeć i zawód)}
\usage{
dane_tab_W1(
  pelna_finalna_ramka_wskaznikow,
  typ_szk,
  kryterium,
  rok_absolwentow,
  rok,
  tylko_tabele
)
}
\arguments{
\item{pelna_finalna_ramka_wskaznikow}{Ramka danych zawierająca pełne wyniki wskaźników.
Oczekuje, że kolumna 'wynik' zawiera zagnieżdżone ramki danych.}

\item{typ_szk}{Zmienna tekstowa opisująca typ szkoły.}

\item{kryterium}{Zmienna tekstowa opisująca kryterium płec(sexf) lub
zawód(nazwa_zaw).}

\item{rok_absolwentow}{Liczba całkowita reprezentująca rok absolwentów do filtrowania.}

\item{rok}{Liczba całkowita reprezentująca rok do filtrowania.}

\item{tylko_tabele}{parametr TRUE/FALSE przekazywany z głównej funkcji
generującej raport - jeśli FALSE to przycina tabele z zawodami do 10
najliczniejszych zawodów, jeśli TRUE to raport generuje tabele zawodów bez
przycinania}
}
\value{
Ramka danych typu tibble w finalnym formacie do zasilania tabel.
}
\description{
Funkcja filtruje i przetwarza dane z pełnej ramki wskaźników,
aby przygotować je do generowania tabel. Tworzy ramkę danych
w formacie będącym wejściem do funkcji gentab_tab_D z tego pakietu.
}
