% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych_K2.R
\name{dane_wyk_K2dzi_plec}
\alias{dane_wyk_K2dzi_plec}
\title{Przygotowanie danych do wykresu kontynuacji nauki (płeć)}
\usage{
dane_wyk_K2dzi_plec(pelna_finalna_ramka_wskaznikow, typ_szk, rok_absolwentow)
}
\arguments{
\item{pelna_finalna_ramka_wskaznikow}{Ramka danych zawierająca pełne wyniki wskaźników.
Oczekuje, że kolumna 'wynik' zawiera zagnieżdżone ramki danych.}

\item{typ_szk}{Zmienna tekstowa opisująca typ szkoły.}

\item{rok_absolwentow}{Liczba całkowita reprezentująca rok absolwentów do filtrowania.}
}
\value{
Ramka danych typu tibble w finalnym formacie do zasilania wykresów.
}
\description{
Funkcja wyciąga i przetwarza dane z pełnej ramki wskaźników,
aby przygotować je do generowania wykresu. Tworzy ramkę danych
z kolumnami  \code{plec}, \code{kontynuacja} oraz \code{pct} (jako odsetek) w formacie
będącym wejściem do funkcji wykresKontynuacjeDziedzinyPlec z pakietu LOSYkolory.
}
