% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych_meta.R
\name{dane_wyk_meta_typsz_zaw}
\alias{dane_wyk_meta_typsz_zaw}
\title{Przygotowanie danych do wykresu zawodów}
\usage{
dane_wyk_meta_typsz_zaw(pelna_finalna_ramka_wskaznikow, rok_absolwentow)
}
\arguments{
\item{pelna_finalna_ramka_wskaznikow}{Ramka danych zawierająca pełne wyniki wskaźników.
Oczekuje, że kolumna 'wynik' zawiera zagnieżdżone ramki danych.}

\item{rok_absolwentow}{Liczba całkowita reprezentująca rok absolwentów do filtrowania.}
}
\value{
Ramka danych typu tibble w finalnym formacie do zasilania wykresów.
}
\description{
Funkcja filtruje dane z pełnej ramki wskaźników,
aby przygotować je do generowania wykresów. Zwraca surową ramkę danych
z zagnieżdżonego obiektu.
}
