% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{dopelniacz}
\alias{dopelniacz}
\title{Generowanie tabeli statusów w formacie flextable}
\usage{
dopelniacz(typ_szkoly)
}
\arguments{
\item{typ_szkoly}{Wartość tekstowa opisująca typ szkoły}
}
\value{
zwraca dopełniach typu szkoły w forie tekstowej
}
\description{
funkcja tworzaca dopełniacz nazwy typu szkoły
}
