% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_tabel_D2.R
\name{gentab_K2dys_zaw}
\alias{gentab_K2dys_zaw}
\title{Generowanie tabeli zawodów w formacie flextable}
\usage{
gentab_K2dys_zaw(dane_do_tabeli, edycja)
}
\arguments{
\item{dane_do_tabeli}{Ramka danych w formacie przygotowanym przez
funkcję \code{dane_wyk_meta_typsz_zaw()}.}

\item{edycja}{Liczba całkowita reprezentująca rok edycji monitoringu.}
}
\value{
Obiekt typu \code{flextable} sformatowany zgodnie z wymaganiami.
}
\description{
Funkcja przyjmuje przygotowane dane dotyczące zawodów i
tworzy z nich sformatowaną tabelę flextable, gotową do wstawienia do raportu Word.
Stosuje styl booktabs, dopasowuje szerokość i dodaje źródło danych.
}
