% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_tabel_D2.R
\name{gentab_tab_D}
\alias{gentab_tab_D}
\title{Generowanie tabeli z danymi z rozdziałów D i K1 w formacie flextable}
\usage{
gentab_tab_D(dane_do_tabeli, edycja)
}
\arguments{
\item{dane_do_tabeli}{Ramka danych w formacie przygotowanym przez funkcje \code{dane_tab_D...}.}

\item{edycja}{Liczba całkowita reprezentująca rok edycji monitoringu.}
}
\value{
Obiekt klasy \code{flextable} sformatowany zgodnie z wymaganiami.
}
\description{
Funkcja przyjmuje przygotowane dane i tworzy z nich sformatowaną tabelę
\code{flextable} do wstawienia do raportu Word. Stosuje styl booktabs, dopasowuje
szerokość i dodaje źródło danych.
}
