% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{\%ni\%}
\alias{\%ni\%}
\title{odwrotność funkcji \code{\%in\%}}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{Wartości do sprawdzenia.}

\item{y}{Wartości, w których należy szukać dopasowania.}
}
\value{
zwraca wektor 1/0
}
\description{
odwraca wektor 0/1 gnerowy przez \code{\%in\%}
}
