% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{tab_num}
\alias{tab_num}
\title{Generowanie kolejnych numerów tabeli}
\usage{
tab_num()
}
\value{
zwraca kolejny nr tabeli
}
\description{
Funkcja do wstawiania kolejnych numerów tabeli w renderowanym
dokumencie
}
