% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{wyk_num}
\alias{wyk_num}
\title{Generowanie kolejnych numer贸w wykres贸w}
\usage{
wyk_num()
}
\value{
zwraca kolejny nr wykresu
}
\description{
Funkcja do wstawiania kolejnych numer贸w wykres贸w w renderowanym
dokumencie
}
